refactor: TripCardStatusLabel & TripCardContent 컴포넌트 분리 [TRL-85]
